Renamed loaders and test classes
