+++ andhow/src/test/java/org/yarnandtail/andhow/AndHow_AliasInTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import static org.yarnandtail.andhow.AndHowTestBase.reloader;
 
-import org.yarnandtail.andhow.load.CmdLineLoader;
+import org.yarnandtail.andhow.load.StringArgumentLoader;
 import org.yarnandtail.andhow.name.BasicNamingStrategy;
 import org.yarnandtail.andhow.property.IntProp;
 import org.yarnandtail.andhow.property.StrProp;
@@ -75,7 +75,7 @@ public class AndHow_AliasInTest extends AndHowTestBase {
 	@Test
 	public void testFirstSetOfInAliasesViaCmdLine() {
 		AndHow.builder().namingStrategy(new BasicNamingStrategy())
-				.loader(new CmdLineLoader())
+				.loader(new StringArgumentLoader())
 				.cmdLineArg(STR_PROP1_IN, STR1)
 				.cmdLineArg(STR_PROP2_ALIAS, STR2)
 				.cmdLineArg(INT_PROP1_ALIAS, INT1.toString())
@@ -91,7 +91,7 @@ public class AndHow_AliasInTest extends AndHowTestBase {
 	@Test
 	public void testSecondSetOfInAliasesViaCmdLine() {
 		AndHow.builder().namingStrategy(new BasicNamingStrategy())
-				.loader(new CmdLineLoader())
+				.loader(new StringArgumentLoader())
 				.cmdLineArg(STR_PROP1_IN_AND_OUT_ALIAS, STR1)
 				.cmdLineArg(STR_PROP2_IN_ALT1_ALIAS, STR2)
 				.cmdLineArg(INT_PROP1_ALT_IN1_ALIAS, INT1.toString())
@@ -106,7 +106,7 @@ public class AndHow_AliasInTest extends AndHowTestBase {
 	@Test
 	public void testThirdSetOfInAliasesViaCmdLine() {
 		AndHow.builder().namingStrategy(new BasicNamingStrategy())
-				.loader(new CmdLineLoader())
+				.loader(new StringArgumentLoader())
 				.cmdLineArg(STR_PROP1_IN_AND_OUT_ALIAS, STR1)
 				.cmdLineArg(STR_PROP2_IN_ALT2_ALIAS, STR2)
 				.cmdLineArg(INT_PROP1_ALT_IN1_ALIAS, INT1.toString())
@@ -218,7 +218,7 @@ public class AndHow_AliasInTest extends AndHowTestBase {
 		
 		try {
 			AndHow.builder().namingStrategy(new BasicNamingStrategy())
-					.loader(new CmdLineLoader())
+					.loader(new StringArgumentLoader())
 					.cmdLineArg(STR_PROP1_IN, STR1)	//minimal values set to ensure no missing value error
 					.cmdLineArg(STR_PROP2_ALIAS, STR2)
 					.cmdLineArg(INT_PROP1_ALIAS, INT1.toString())
@@ -245,7 +245,7 @@ public class AndHow_AliasInTest extends AndHowTestBase {
 		
 		try {
 			AndHow.builder().namingStrategy(new BasicNamingStrategy())
-					.loader(new CmdLineLoader())
+					.loader(new StringArgumentLoader())
 					.cmdLineArg(STR_PROP1_IN, STR1)	//minimal values set to ensure no missing value error
 					.cmdLineArg(STR_PROP2_ALIAS, STR2)
 					.cmdLineArg(INT_PROP1_ALIAS, INT1.toString())
@@ -271,7 +271,7 @@ public class AndHow_AliasInTest extends AndHowTestBase {
 		
 		try {
 			AndHow.builder().namingStrategy(new BasicNamingStrategy())
-					.loader(new CmdLineLoader())
+					.loader(new StringArgumentLoader())
 					.cmdLineArg(STR_PROP1_IN, STR1)	//minimal values set to ensure no missing value error
 					.cmdLineArg(STR_PROP2_ALIAS, STR2)
 					.cmdLineArg(INT_PROP1_ALIAS, INT1.toString())
