+++ andhow/src/test/java/org/yarnandtail/andhow/load/PropertyFileOnClasspathLoaderAppTest.java
@@ -3,11 +3,8 @@ package org.yarnandtail.andhow.load;
 import org.yarnandtail.andhow.AndHow;
 import org.yarnandtail.andhow.AppFatalException;
 import org.yarnandtail.andhow.PropertyGroup;
-import java.io.File;
-import java.net.URL;
 import java.util.List;
-import org.apache.commons.io.FileUtils;
-import org.junit.*;
+import org.junit.Test;
 import org.yarnandtail.andhow.*;
 
 import static org.junit.Assert.*;
@@ -24,46 +21,25 @@ import static org.yarnandtail.andhow.AndHowTestBase.reloader;
  * some of the higher-level errors can be tested
  * @author eeverman
  */
-public class PropertyFileFromFilesystemLoaderAppTest {
-
-	File tempPropertiesFile = null;
+public class PropertyFileOnClasspathLoaderAppTest {
 	
 	public static interface TestProps extends PropertyGroup {
-		StrProp FILEPATH = StrProp.builder().build();
-	}
-	
-
-	@Before
-	public void init() throws Exception {
-		
-		//copy a properties file to a temp location
-		URL inputUrl = getClass().getResource("/org/yarnandtail/andhow/load/SimpleParams1.properties");
-		tempPropertiesFile = File.createTempFile("andhow_test", ".properties");
-		tempPropertiesFile.deleteOnExit();
-		FileUtils.copyURLToFile(inputUrl, tempPropertiesFile);
-
-	}
-	
-	@After
-	public void afterTest() {
-		if (tempPropertiesFile != null) {
-			tempPropertiesFile.delete();
-		}
+		StrProp CLAZZ_PATH = StrProp.builder().build();
 	}
 	
 	@Test
 	public void testHappyPath() throws Exception {
 		AndHow.builder().namingStrategy(new BasicNamingStrategy())
-				.loader(new CmdLineLoader())
-				.loader(new PropertyFileFromFilesystemLoader(TestProps.FILEPATH))
-				.cmdLineArg(PropertyGroup.getCanonicalName(TestProps.class, TestProps.FILEPATH), 
-						tempPropertiesFile.getAbsolutePath())
+				.loader(new StringArgumentLoader())
+				.loader(new PropertyFileOnClasspathLoader(TestProps.CLAZZ_PATH))
+				.cmdLineArg(PropertyGroup.getCanonicalName(TestProps.class, TestProps.CLAZZ_PATH), 
+						"/org/yarnandtail/andhow/load/SimpleParams1.properties")
 				.group(SimpleParams.class)
 				.group(TestProps.class)
 				.reloadForNonPropduction(reloader);
 		
 
-		assertEquals(tempPropertiesFile.getAbsolutePath(), TestProps.FILEPATH.getValue());
+		assertEquals("/org/yarnandtail/andhow/load/SimpleParams1.properties", TestProps.CLAZZ_PATH.getValue());
 		assertEquals("kvpBobValue", SimpleParams.STR_BOB.getValue());
 		assertEquals("kvpNullValue", SimpleParams.STR_NULL.getValue());
 		assertEquals(Boolean.FALSE, SimpleParams.FLAG_TRUE.getValue());
@@ -76,10 +52,10 @@ public class PropertyFileFromFilesystemLoaderAppTest {
 		
 		try {
 			AndHow.builder().namingStrategy(new BasicNamingStrategy())
-					.loader(new CmdLineLoader())
-					.loader(new PropertyFileFromFilesystemLoader(TestProps.FILEPATH))
-					.cmdLineArg(PropertyGroup.getCanonicalName(TestProps.class, TestProps.FILEPATH), 
-							tempPropertiesFile.getAbsolutePath())
+					.loader(new StringArgumentLoader())
+					.loader(new PropertyFileOnClasspathLoader(TestProps.CLAZZ_PATH))
+					.cmdLineArg(PropertyGroup.getCanonicalName(TestProps.class, TestProps.CLAZZ_PATH), 
+							"/org/yarnandtail/andhow/load/SimpleParams1.properties")
 					.group(SimpleParams.class)
 					//.group(TestProps.class)	//This must be declared or the Prop loader can't work
 					.reloadForNonPropduction(reloader);
@@ -93,7 +69,7 @@ public class PropertyFileFromFilesystemLoaderAppTest {
 	}
 	
 	/**
-	 * It is not an error to not specify the file path property, it just means the loader
+	 * It is not an error to not specify the classpath param, it just means the laoder
 	 * will not find anything.
 	 * 
 	 * @throws Exception 
@@ -101,8 +77,8 @@ public class PropertyFileFromFilesystemLoaderAppTest {
 	@Test
 	public void testUnspecifiedConfigParam() throws Exception {
 		AndHow.builder().namingStrategy(new BasicNamingStrategy())
-				.loader(new CmdLineLoader())
-				.loader(new PropertyFileFromFilesystemLoader(TestProps.FILEPATH))
+				.loader(new StringArgumentLoader())
+				.loader(new PropertyFileOnClasspathLoader(TestProps.CLAZZ_PATH))
 				.group(SimpleParams.class)
 				.group(TestProps.class)
 				.reloadForNonPropduction(reloader);
@@ -117,9 +93,9 @@ public class PropertyFileFromFilesystemLoaderAppTest {
 		
 		try {
 			AndHow.builder().namingStrategy(new BasicNamingStrategy())
-					.loader(new CmdLineLoader())
-					.loader(new PropertyFileFromFilesystemLoader(TestProps.FILEPATH))
-					.cmdLineArg(PropertyGroup.getCanonicalName(TestProps.class, TestProps.FILEPATH), 
+					.loader(new StringArgumentLoader())
+					.loader(new PropertyFileOnClasspathLoader(TestProps.CLAZZ_PATH))
+					.cmdLineArg(PropertyGroup.getCanonicalName(TestProps.class, TestProps.CLAZZ_PATH), 
 							"asdfasdfasdf/asdfasdf/asdf")
 					.group(SimpleParams.class)
 					.group(TestProps.class)
